+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Changelog for v0.2.0
+
+## Changes
+
+* `Interceptor.Configurator` provides a DSL to define the intercept configuration, allowing `"Module.function/arity"` MFAs instead of tuple-based ones
+* Intercept configuration can now live directly on the intercepted module, instead of being exposed by a module set on the application configuration
+* Ability to intercept private functions as well
+* Refactor of the configuration code to its own `Interceptor.Configuration` module
+
+## TODO
+
+* Allow multiple callbacks for a given moment, i.e., allow more than one callback to be invoked `after` the intercepted function is called, for example.
+
 # Changelog for v0.1.3
 
 ## Changes